Fjernet muligheten for å endre til desimaler fra frontend

+++ bff-nav_ansatt/src/main/kotlin/no/nav/amt/tiltak/bff/nav_ansatt/dto/EndringsmeldingDto.kt
@@ -53,7 +53,7 @@ data class EndringsmeldingDto(
 
 		data class EndreDeltakelseProsentInnhold(
 			val deltakelseProsent: Int,
-			val dagerPerUke: Float?,
+			val dagerPerUke: Int?,
 			val gyldigFraDato: LocalDate?
 		) : Innhold()
