+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/endringsmelding/EndringsmeldingDbo.kt
@@ -54,7 +54,7 @@ data class EndringsmeldingDbo(
 
 		data class EndreDeltakelseProsentInnhold(
 			val nyDeltakelseProsent: Int,
-			val dagerPerUke: Float?,
+			val dagerPerUke: Int?,
 			val gyldigFraDato: LocalDate?
 		): Innhold()
